services/service-ai-analyzer/requirements.txt
pika
pydantic
pydantic-settings
google-generativeai
instructor>=1.6.0
